Move test to a mod test and add more resilient test logic
